Detail Update

+++ app/src/main/java/com/kotlin/andi/mymovies/domain/model/Movie.kt
@@ -15,7 +15,7 @@ data class Movie(
     val backdropPath: String? = null,
     val releaseDate: String? = null,
     val popularity: Double? = null,
-    val voteAverage: Double? = null,
+    val voteAverage: Float? = null,
     val id: Int? = null,
     val adult: Boolean? = null,
     val voteCount: Int? = null,
